requirements.txt
google-genai
dotenv
fastapi 
pydantic[email]
uvicorn
gunicorn